refactor: context and components
