+++ src/components/index.ts
@@ -1,0 +1,6 @@
+
+
+export { FavoriteWords } from "@/components/FavoriteWords";
+export { FormSearchWord } from "@/components/FormSearchWord";
+export { SingleWordPanel } from "@/components/SingleWordPanel";
+export { WordInformation } from "@/components/WordInformation";
